# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FireTrade Mobile Application API
  version: 1.0.0
  description: |
    MVP Backend API for FireTrade - A Crypto Exchange and Bill Payment Service
    supporting NGN fiat pairs and digital bill payments in Nigeria.
  contact:
    name: FireTrade API Support
    email: api@firetrade.ng
servers:
  - url: http://localhost:3002/api/v1/
    description: Production server
  - url: https://staging-api.firetrade.ng/api/v1
    description: Staging server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token obtained from authentication endpoints
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - phone
        - full_name
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePass123!"
        phone:
          type: string
          pattern: '^\+234[0-9]{10}$'
          example: "+2348012345678"
        full_name:
          type: string
          minLength: 3
          example: "John Doe"
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          example: "user@example.com"
        requires_2fa_setup:
          type: boolean
          example: true
        qr_code_url:
          type: string
          example: "otpauth://totp/FireTrade:user@example.com?secret=..."
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        requires_2fa:
          type: boolean
          example: true
        expires_in:
          type: integer
          example: 3600
    TwoFAVerifyRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          pattern: "^[0-9]{6}$"
          example: "123456"
    TwoFAVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
          example: true
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    BillCategory:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
          example: "a50e8400-e29b-41d4-a716-446655440007"
        name:
          type: string
          example: "Airtime"
        description:
          type: string
          example: "Mobile airtime recharge"
        providers:
          type: array
          items:
            type: object
            properties:
              provider_id:
                type: string
                example: "MTN"
              name:
                type: string
                example: "MTN Nigeria"
              min_amount:
                type: string
                example: "50.00"
              max_amount:
                type: string
                example: "50000.00"
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with mandatory KYC information
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user credentials and returns JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/2fa/verify:
    post:
      tags:
        - Authentication
      summary: Verify 2FA code
      description: Validates the 2FA TOTP code during login or setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAVerifyRequest"
      responses:
        "200":
          description: 2FA verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFAVerifyResponse"
        "401":
          description: Invalid 2FA code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many verification attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bills/categories:
    get:
      tags:
        - Bill Payments
      summary: Get bill payment categories
      description: Retrieves available bill payment categories (Airtime, Data, etc.)
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/BillCategory"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...